refactor: share toggleItem util
