+++ packages/shared-utils/src/index.ts
@@ -1,3 +1,4 @@
 export { fetchJson } from './fetchJson';
 export { slugify } from './slugify';
 export { genSecret } from './genSecret';
+export { toggleItem } from './toggleItem';
