+++ packages/ui/src/components/cms/PublishLocationSelector.tsx
@@ -4,7 +4,7 @@
 import { Button, Input } from "@ui/components/atoms/shadcn";
 import type { PublishLocation } from "@types";
 import { usePublishLocations } from "@ui/hooks/usePublishLocations";
-import { toggleItem } from "../../utils";
+import { toggleItem } from "@shared-utils";
 import { memo, useCallback } from "react";
 
 export interface PublishLocationSelectorProps {
