+++ packages/ui/src/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './style';
 export * from './i18n';
-export * from './toggleItem';
